Rename {InConfiguration -> ConfigurationCacheable}IdFactory
