+++ subprojects/core/src/main/java/org/gradle/internal/buildtree/BuildTreeScopeServices.java
@@ -27,7 +27,7 @@ import org.gradle.api.logging.configuration.LoggingConfiguration;
 import org.gradle.api.logging.configuration.ShowStacktrace;
 import org.gradle.cache.internal.DecompressionCacheFactory;
 import org.gradle.cache.scopes.BuildTreeScopedCacheBuilderFactory;
-import org.gradle.configuration.internal.InConfigurationIdFactory;
+import org.gradle.internal.id.ConfigurationCacheableIdFactory;
 import org.gradle.execution.DefaultTaskSelector;
 import org.gradle.execution.ProjectConfigurer;
 import org.gradle.execution.TaskNameResolver;
@@ -85,7 +85,7 @@ public class BuildTreeScopeServices {
         registration.add(DefaultProblemLocationAnalyzer.class);
         registration.add(DefaultExceptionAnalyser.class);
         registration.add(ScriptUsageLocationReporter.class);
-        registration.add(InConfigurationIdFactory.class);
+        registration.add(ConfigurationCacheableIdFactory.class);
         registration.add(TaskIdentityFactory.class);
         modelServices.applyServicesTo(registration);
     }
